Refactored event loop into an struct
The passing of parameters in big chains produces not very readable code,
so refactoring most functions into getting most of their state from a
overarching struct seems reasonable. With this, we aliminated most
parameters in most functions and we are able to make changes easier in
both functions and in the struct.

The code follows the same flow as before.